extract psi elems from file

+++ src/main/kotlin/Fragment.kt
@@ -1,2 +1,2 @@
-class Fragment(var file: String, var line : Long, var start : Long, var end : Long) {
+class Fragment(var file: String, var line : Long, var start : Int, var end : Int) {
 }
